requirements.txt
attrs
flake8
mypy
pydocstyle
pygame
pyproject-flake8
PyYAML
